# Performance & other misc patches