This pull request will address the following issues,

**Dependency Issue**

1. Reduce guava dependency version to 10.0.1 to match Spark’s dependency

**Functionality Issues**

1. Error in the sample program where data generation is done
2. In multi GPU mode, Caching flag is not set in the executors 
3. In multi GPU mode, executors don’t get assign to all GPUs
4. In multi GPU mode, Cached Dataset list is not passed on to newly created executors
5. Introduce concurrent hash maps to avoid deadlocks by threads operating on different partitions on an executor

**Performance Isses**

1. CPU vs GPU performance is hit and the following patches are made to improve it,
	a. GPU pointers caching should be done at individual partitions to gain runtime performance
	b. Caching the partition size for the cached datasets will help gain performance by eliminating loops

Performance Numbers for Logistic Regression(10 iteration) is as follows,

```
============ GPU =======================
Iteration 1 Done in 1627 ms
Iteration 2 Done in 1669 ms
Iteration 3 Done in 1512 ms
Iteration 4 Done in 1563 ms
Iteration 5 Done in 1469 ms
Iteration 6 Done in 1502 ms
Iteration 7 Done in 1543 ms
Iteration 8 Done in 1436 ms
Iteration 9 Done in 1517 ms
Iteration 10 Done in 1489 ms
GPU Elapsed time: 15333 ms

============ CPU =======================
Iteration 1 Done in 7349 ms
Iteration 2 Done in 5532 ms
Iteration 3 Done in 5114 ms
Iteration 4 Done in 5002 ms
Iteration 5 Done in 5367 ms
Iteration 6 Done in 4919 ms
Iteration 7 Done in 4655 ms
Iteration 8 Done in 4757 ms
Iteration 9 Done in 4498 ms
Iteration 10 Done in 4603 ms
CPU Elapsed time: 51803 ms
```