@josiahsams 2017-06-09
Performance number for Logistic Regression when running it in a Single Node Multi-GPU(4 NVidia Tesla K80) environment with 4 Million Rows & 400 features,
```
$ time ~/spark/bin/spark-submit 
 --master spark://soe15:7077 
 --class com.ibm.gpuenabler.SparkDSLR  
 --jars ~/new/GPUEnabler/gpu-enabler/target/gpu-enabler_2.11-1.0.0.jar  
~/new/GPUEnabler/examples/target/gpu-enabler-examples_2.11-1.0.0.jar 
  spark://soe15:7077 16 4000000 400 10

Iteration 1 Done in 2486 ms
Iteration 2 Done in 2194 ms
Iteration 3 Done in 2765 ms
Iteration 4 Done in 1976 ms
Iteration 5 Done in 1933 ms
Iteration 6 Done in 2032 ms
Iteration 7 Done in 2079 ms
Iteration 8 Done in 1769 ms
Iteration 9 Done in 1974 ms
Iteration 10 Done in 2091 ms

GPU Elapsed time: 21306 ms
```